Fix error with ignore 404 in Delete bill run (#1)
We added a condition that meant the `404` returned when the bill run status `GET` call is used is not seen by JMeter as an error. What we weren't able to confirm at the time is if it returns a `200` would it still work (it doesn't!) We just didn't have enough data to cause the delete process to take long enough for polling to be needed.

This change fixes the issue by updating the condition in the [JMeter response assetion](https://jmeter.apache.org/usermanual/component_reference.html#Response_Assertion) to check for both `200` or `404`.