Add delete licence journey
https://eaflood.atlassian.net/browse/CMEA-141

WRLS highlighted an issue with deleting a licence. Because the endpoint immediately responds with a `204` and does the delete in the background it's not obvious when the task is complete and the bill run has been updated.

So, we applied a fix where when the delete starts it marks the bill run as `pending` and then resets it back when finished. But we wanted to confirm this happened through some form of test and JMeter seemed like a good place to start. We need to generate enough transactions that the delete takes a bit of time. We also need to be able to poll the bill run status and our JMeter tests already have examples of doing this.

Hence this change, which adds a delete licence journey to our performance tests.

> In the end we ended up updating our test bill run generator endpoint which means we can generate a large number of transactions very quickly from one request. So, in future it might be easier to use our new [https://github.com/DEFRA/sroc-charging-module-api-tests](https://github.com/DEFRA/sroc-charging-module-api-tests) to check for something like this.