Add new thread count arguments
We need to add one for each test (thread group) we have in the test plan (`tests.jmx`). With this we just need to set the value to 1 or more and the thread group will be run.

As standard journey is our main test we have that set to 1.

We also update the JMX file such that all the listeners on 'Standard Journey' are disabled. This is recommended way to run a test via the command in order to maximise performance and get the most accurate result.

There is nothing stopping a user re-enabling them when in GUI mode. Just a shame there isn't a CLI arg to disable all listeners.