Ignore the JMeter test output file
We'll want some output when we run in CLI mode. JMeter can output a table of the results for each request made known as a .jtl file (it's just a CSV file!)

We'll have JMeter generate them when we run the CLI version but we don't want them committed to the project.