Add Docker support to the project
One day, we'd like to integrate the performance tests into our CI. We'd also like to make it easier to run the tests. All members of the current team had to battle to get JMeter running on their machines. Those actively working on the tests will still need to work as normal in order to run JMeter in GUI mode. But those who just want to run the tests will have an easier time of it if we can dockerise the project.

Building on [Better support running tests using JMeter CLI](https://github.com/DEFRA/sroc-cha-performance-tests/pull/3) this change add's the ability to build a Docker image that contains the project and ensures we can run our tests from within it.