Better support running tests using JMeter CLI (#3)
Whenever we start JMeter we get this message

```
================================================================================
Don't use GUI mode for load testing !, only for Test creation and Test debugging.
For load testing, use CLI Mode (was NON GUI):
   jmeter -n -t [jmx file] -l [results file] -e -o [Path to web report folder]
& increase Java Heap to meet your test requirements:
   Modify current env variable HEAP="-Xms1g -Xmx1g -XX:MaxMetaspaceSize=256m" in the jmeter batch file
Check : https://jmeter.apache.org/usermanual/best-practices.html
================================================================================
```

The problem is we don't want to split out our JMX file as we often find we need to copy something from one Thread Group to another (we know there are ways of breaking up JMeter tests but there is only so much time in the day!)

Launching JMeter in GUI mode means we can easily control which Thread Group will get run by disabling/enabling where necessary. But we have a desire to also add Docker support to the project with an eye on building the tests into our CI and for that, we'll need to use the command line.

There isn't a way to specify a specific thread group via the command line. But we can take advantage of a common method which is to set the `Number of Threads (users)` to 0 for those we don't want to run and 1 or greater for those we do. To do this we'll need to add a new env var for each group (test). We also add an alternate to `run.sh` which will run using the CLI.

We can then take advantage of this locally (and shut JMeter up 😀). But it will also help us when we come to dockerise the project.