Add VSCode helper tasks
All the current team are VSCode users and we have found it helpful in projects that use Docker to add helper tasks. Trying to remember the complex commands needed to build and run docker containers is not fun. So, by using VSCode tasks we get 2 wins; we document the commands needed and we add a mechanism that allows us to use them from within our preferred editor!